Add ERRC20 historical balances
